requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.3
cryptography==44.0.3
cycler==0.12.1
datasets==3.1.0
dill==0.3.8
filelock==3.20.0
flwr==1.20.0
flwr-datasets==0.5.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.9.0
grpcio==1.76.0
grpcio-health-checking==1.62.3
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.11
iterators==0.0.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
networkx==3.5
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
propcache==0.4.1
protobuf==4.25.8
pyarrow==21.0.0
pycparser==2.23
pycryptodome==3.23.0
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
ray==2.31.0
referencing==0.37.0
requests==2.32.5
rich==13.9.4
rpds-py==0.27.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tomli==2.3.0
tomli_w==1.2.0
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
typer==0.12.5
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.6.0
yarl==1.22.0
